move constructWebHookData to webhook invoker
